Added support for users not being logged in

+++ src/lib/index.ts
@@ -1,6 +1,6 @@
 import { browser } from '$app/environment';
 import type { tasks } from '$lib/server/db/schema';
-import type { Task } from '$lib/types';
+import { type Task, TaskStatus } from '$lib/types';
 import { tasker } from '$lib/state.svelte';
 
 export function playSound(type: 'task-started' | 'reminder') {
@@ -71,42 +71,64 @@ export function playSound(type: 'task-started' | 'reminder') {
 }
 
 export async function createTask(taskData: typeof tasks.$inferInsert) {
-	const response = await fetch('/edit', {
-		method: 'POST',
-		headers: {
-			'Content-Type': 'application/json'
-		},
-		body: JSON.stringify(taskData)
-	});
+	if (taskData.email) {
+		const response = await fetch('/edit', {
+			method: 'POST',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify(taskData)
+		});
 
-	if (!response.ok) {
-		throw new Error('Failed to create task');
+		if (!response.ok) {
+			throw new Error('Failed to create task');
+		}
+
+		tasker.tasks = [...tasker.tasks, await response.json()];
+	} else {
+		// If no email is provided, add to localStorage
+		tasker.tasks = [
+			...tasker.tasks,
+			{
+				...taskData,
+				status: TaskStatus.Pending,
+				id: Date.now()
+			} as Task
+		];
+		saveTasksLocally();
 	}
-
-	tasker.tasks = [...tasker.tasks, await response.json()];
 }
 
 export async function updateTask(taskData: Task) {
-	await fetch('/edit', {
-		method: 'PUT',
-		headers: {
-			'Content-Type': 'application/json'
-		},
-		body: JSON.stringify(taskData)
-	});
-
 	const taskIndex = tasker.tasks.findIndex((t) => t.id === taskData.id);
 	if (taskIndex === -1) return;
 	tasker.tasks[taskIndex] = { ...taskData };
+
+	if (taskData.email) {
+		await fetch('/edit', {
+			method: 'PUT',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify(taskData)
+		});
+	} else saveTasksLocally();
 }
 
 export async function deleteTask(id: number) {
-	await fetch('/edit', {
-		method: 'DELETE',
-		headers: {
-			'Content-Type': 'application/json'
-		},
-		body: JSON.stringify({ id })
-	});
+	const task: Task = tasker.tasks.filter((task) => task.id === id)[0];
 	tasker.tasks = tasker.tasks.filter((task) => task.id !== id);
+	if (task.email) {
+		await fetch('/edit', {
+			method: 'DELETE',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify({ id })
+		});
+	} else saveTasksLocally();
+}
+
+function saveTasksLocally() {
+	localStorage.setItem('tasks', JSON.stringify(tasker.tasks));
 }
